+++ packages/@dicebear/collection/src/index.ts
@@ -1,12 +1,9 @@
 export * as avataaars from '@dicebear/avatars-avataaars-sprites';
 export * as bottts from '@dicebear/avatars-bottts-sprites';
-export * as female from '@dicebear/avatars-female-sprites';
 export * as gridy from '@dicebear/avatars-gridy-sprites';
-export * as human from '@dicebear/avatars-human-sprites';
-export * as identicon from '@dicebear/avatars-identicon-sprites';
+export * as identicon from '@dicebear/identicon';
 export * as initials from '@dicebear/initials';
 export * as jdenticon from '@dicebear/avatars-jdenticon-sprites';
-export * as male from '@dicebear/avatars-male-sprites';
 export * as micah from '@dicebear/micah';
 export * as croodles from '@dicebear/croodles';
 export * as croodlesNeutral from '@dicebear/croodles-neutral';