+++ packages/@dicebear/identicon/src/index.ts
@@ -1,0 +1,15 @@
+/*!
+ * Identicon (@dicebear/identicon)
+ *
+ * Code licensed under MIT License.
+ * Copyright (c) 2021 Florian Körner
+ *
+ * Design "Identicon".
+ */
+
+import { style } from './core';
+
+let { create, meta, schema } = style;
+
+export { create, meta, schema };
+export { Options } from './options';