+++ plugins/figma/dicebear-exporter/src/ui/components/LoadedScene.svelte
@@ -3,7 +3,7 @@
   import { activeStage } from '../stores/activeStage';
 
   import MenuItem from './MenuItem.svelte';
-  import BackgroundForm from './BackgroundForm.svelte';
+  import GeneralForm from './GeneralForm.svelte';
   import PackageForm from './PackageForm.svelte';
   import LicenseForm from './LicenseForm.svelte';
   import HookForm from './HookForm.svelte';
@@ -29,7 +29,7 @@
     <div class="menu-section">Frame</div>
     <MenuItem stage={'package'}>Package</MenuItem>
     <MenuItem stage={'license'}>License</MenuItem>
-    <MenuItem stage={'background'}>Background</MenuItem>
+    <MenuItem stage={'general'}>General</MenuItem>
     <MenuItem stage={'alias'}>Aliases</MenuItem>
     <MenuItem stage={'hook'}>Hooks</MenuItem>
   </div>
@@ -64,8 +64,8 @@
       <PackageForm />
     {:else if activeStageSplit[0] === 'license'}
       <LicenseForm />
-    {:else if activeStageSplit[0] === 'background'}
-      <BackgroundForm />
+    {:else if activeStageSplit[0] === 'general'}
+      <GeneralForm />
     {:else if activeStageSplit[0] === 'hook'}
       <HookForm />
     {:else if activeStageSplit[0] === 'alias'}